🤖 Chore: move channel id
